[util] Remove parse_util.h
 - Deleted parse_util.h, parse_util.cpp
 - Move functions in props_util.h to util_func.h and deleted props_util.h
 - Remove unused function in util_func

Self evaluation:

Build test: [X]Passed [ ]Failed [ ]Skipped
Run test: [X]Passed [ ]Failed [ ]Skipped
