+++ nntrainer/compiler/ini_interpreter.cpp
@@ -21,7 +21,6 @@
 #include <layer_node.h>
 #include <nntrainer_error.h>
 #include <nntrainer_log.h>
-#include <parse_util.h>
 #include <time_dist.h>
 #include <util_func.h>
