+++ nntrainer/layers/activation_layer.cpp
@@ -25,7 +25,6 @@
 #include <nntrainer_error.h>
 #include <nntrainer_log.h>
 #include <node_exporter.h>
-#include <parse_util.h>
 #include <tensor.h>
 #include <util_func.h>
