+++ nntrainer/layers/conv2d_layer.cpp
@@ -22,7 +22,6 @@
 #include <nntrainer_error.h>
 #include <nntrainer_log.h>
 #include <node_exporter.h>
-#include <parse_util.h>
 #include <profiler.h>
 #include <util_func.h>
