+++ nntrainer/layers/layer_impl.cpp
@@ -19,7 +19,6 @@
 #include <nntrainer_error.h>
 #include <nntrainer_log.h>
 #include <node_exporter.h>
-#include <parse_util.h>
 #include <util_func.h>
 namespace nntrainer {
