+++ nntrainer/layers/layer_node.cpp
@@ -18,7 +18,6 @@
 #include <nntrainer_error.h>
 #include <nntrainer_log.h>
 #include <node_exporter.h>
-#include <parse_util.h>
 #include <time_dist.h>
 #include <util_func.h>
