+++ nntrainer/tensor/tensor_dim.cpp
@@ -19,8 +19,8 @@
 
 #include <nntrainer_error.h>
 #include <nntrainer_log.h>
-#include <parse_util.h>
 #include <tensor_dim.h>
+#include <util_func.h>
 
 namespace ml {
 namespace train {
