+++ nntrainer/utils/base_properties.cpp
@@ -10,7 +10,6 @@
  * @bug No known bugs except for NYI items
  */
 #include <base_properties.h>
-#include <parse_util.h>
 
 #include <regex>
 #include <sstream>
